refactor: split biped silhouette definitions
